Move StackArray to edu.kis.vh.stacks.list package.
