+++ src/main/java/edu/kis/vh/stacks/list/StackArray.java
@@ -1,6 +1,4 @@
-package edu.kis.vh.stacks;
-
-import edu.kis.vh.stacks.list.Stackable;
+package edu.kis.vh.stacks.list;
 
 public class StackArray implements Stackable {
 	/**
